[#173] cleanup event api

+++ examples/examples/event/notifier.rs
@@ -24,7 +24,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let notifier = event.notifier().create()?;
 
-    let mut counter: u64 = 0;
+    let mut counter: usize = 0;
     while let Iox2Event::Tick = Iox2::wait(CYCLE_TIME) {
         counter += 1;
         notifier.notify_with_custom_event_id(EventId::new(counter))?;
